requirements.txt
apache-airflow==2.2.4
pandas==1.1.5
snscrape==0.3.4
pymongo==3.12.3
boto3==1.21.30
awscli==1.22.85
nltk==3.6.7
gensim==4.1.2